change the naming and logic of the consistency check again
I scrutinised the kvalobs code, and it's not very consistent, but this
best matches the behaviour in a consistent way.

The only thing left to consider is whether we should return separate
flags for the sequence values. kvalobs flags only one of them (and is
rather inconsistent about which one some flag the min/max while others
flag only the first value weirdly). I will consult with the others in
Confident WP5 and revisit this based on feedback.